# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이것은 Project. Stock Exchange API 문서입니다.",
        "title": "Project. Stock Exchange API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/0ghost0-dev/PJSe/blob/master/license"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/": {
            "get": {
                "description": "유효한 계정으로 인증된 경우 \"Authenticated\" 메시지와 유저 정보를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "인증 테스트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Basic {BASE64_ENCODED_CREDENTIALS}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 인증 메시지 및 유저 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "새로운 유저를 등록합니다. (Access Code 필요)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "유저 등록",
                "parameters": [
                    {
                        "description": "유저 등록 정보",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "성공 시 유저 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/api": {
            "get": {
                "description": "API 키 상태 확인",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "API 키 상태 확인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 API 키 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/auth/token": {
            "post": {
                "description": "인증된 유저를 위해 새로운 API 키를 생성합니다. (임시 키, 24시간 유효)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "임시 API 키 생성",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Basic {base64_encoded_credentials}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 새로운 API 키 및 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/market/status": {
            "get": {
                "description": "거래소의 현재 세션 상태(오픈, 클로즈 등)를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market - Status"
                ],
                "summary": "거래소 세션 정보 조회",
                "responses": {
                    "200": {
                        "description": "성공 시 세션 상태 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/market/status/exchange-data": {
            "get": {
                "description": "거래소의 현재 데이터(심볼, 티커 등)를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market - Status"
                ],
                "summary": "거래소 데이터 조회",
                "responses": {
                    "200": {
                        "description": "성공 시 거래소 데이터 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}